requirements.txt
attrs==22.2.0